pkg/gadgettracermanager: Implement a Pod informer
Inspektor Gadget has to perform some operations before a container is created
to trace it. The OCI hooks provide a mechanishm to be notified when a container
is created, unfortunately it's not possible to use OCI hooks in all the cases
as they are not supported by all the container runtime interfaces, cri-o is the
only one with a propper support, containerd support is on the roadmap and
docker doesn't have any plans.

This commit adds an alternative based on a controller that watches for pods.
This solution is not an ideal because the nofication received from the k8s API
server is not synchronous, i.e. the nofication arrives after the container
is created and there is not any way to pause the creation while Inspektor Gadget
performs the operations to be ready to trace that container. It implies that
IG could lose the first events on the container.